app_segmentation/requirements.txt

grpcio==1.42.0
icecream==2.1.3
numpy==1.21.5
Pillow==9.4.0
protobuf==4.22.1
scikit_image==0.19.2